:art: Add upload CLI command
